Added FORTRAN apt-get
I was hiting errors with the discovery of the FORTRAN compiler R needs on import via RPy.  Rather than building things locally and replumbing a lot of infrastructure, this is a dumb hack to try to get the files up there so I can us LD_LIBRARY_PATH to discover them.